pyproject.toml
[tool.poetry]
name = "ontologies-explorer"
version = "0.1.0"
description = ""
authors = [
    "lnmylos <elenamylonas@gmail.com>",
    "Goldensteev <fletcher.steven@hotmail.fr>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
PyYAML = "^6.0.1"
pydantic = { extras = ["email"], version = "^2.6.3" }
pytest = "^8.1.1"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.8.0"
pylint = "^3.1.0"
ruff = "^0.5.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]


line-length = 100
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]


[tool.pytest.ini_options]
addopts = "-v --cov --cov-report=term --cov-report=xml"

[tool.coverage.run]
source = ["utils"]
omit = ["*/.vscode/extensions/*", "tests/*"]
